moodle-admin-app/requirements.txt
Flask==3.1.0
PyMySQL==1.1.1
bcrypt==4.2.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
pytest==7.4.0
